Refactor process management; implement getPid method in SystemProcess and update ProcessMonitor to retrieve PID correctly
