+++ src/maincontroller/main_controller.cpp
@@ -3,27 +3,9 @@
 #include <thread>
 #include "console_control.h"
 #include "nap_time.h"
-#include "network_process.h"
-#include "process.h"
-#include "process_simulator.h"
-#include "system_monitor.h"
-#include "system_process.h"
 
 namespace process
 {
-    // Define the handler factory map
-    FactoryMap MainController::handlerFactoryMap_;
-
-    std::atomic<int> MainController::counter_ = 0;
-    // Initialize the factory map
-    void MainController::initializeFactory()
-    {
-        handlerFactoryMap_["real"]    = []() { return manufacture<ProcessMonitor, Process>(); };
-        handlerFactoryMap_["simul"]   = []() { return manufacture<ProcessMonitor, ProcessSimulator>(); };
-        handlerFactoryMap_["network"] = []() { return manufacture<ProcessMonitor, NetworkProcess>(); };
-        handlerFactoryMap_["system"]  = []() { return manufacture<SystemMonitor, SystemProcess>(); };
-    }
-
     void MainController::run(const std::string &processType, int numProcesses)
     {
         ProcessController::setProcessType(processType);
