+++ src/maincontroller/main_controller_createthreads.cpp
@@ -9,9 +9,26 @@
 #include "string_tools.h"
 #include "system_monitor.h"
 #include "ProcessRegistry.h"
+#include "network_process.h"
+#include "process.h"
+#include "process_simulator.h"
+#include "system_process.h"
 
 namespace process
 {
+    // Define the handler factory map
+    FactoryMap MainController::handlerFactoryMap_;
+
+    std::atomic<int> MainController::counter_ = 0;
+    // Initialize the factory map
+    void MainController::initializeFactory()
+    {
+        handlerFactoryMap_["real"]    = []() { return manufacture<ProcessMonitor, Process>(); };
+        handlerFactoryMap_["simul"]   = []() { return manufacture<ProcessMonitor, ProcessSimulator>(); };
+        handlerFactoryMap_["network"] = []() { return manufacture<ProcessMonitor, NetworkProcess>(); };
+        handlerFactoryMap_["system"]  = []() { return manufacture<SystemMonitor, SystemProcess>(); };
+    }
+
     void MainController::restoreHandlerCount()
     {
         // Check if the number of handlers is less than numProcesses_
