+++ src/processmonitors/system_monitor.cpp
@@ -1,11 +1,11 @@
 #include "system_monitor.h"
-#include <unistd.h>
 #include "logger_instance.h"
 
 namespace process
 {
     void SystemMonitor::launchChildProcess() 
     { 
-        process_->work(); 
+        process_->work();
+        this->pid_ = process_->getPid();
     }
 }
